fix: 修改旧的type命名为model
