+++ src/main/java/com/simonalong/mikilin/annotation/FieldWhiteMatcher.java
@@ -36,7 +36,7 @@ public @interface FieldWhiteMatcher {
      * 可用的值对应的类型
      * @return 对应的枚举类型
      */
-    FieldModel type() default FieldModel.DEFAULT;
+    FieldModel model() default FieldModel.DEFAULT;
 
     /**
      * 枚举类型的判断
