+++ src/test/java/com/simonalong/mikilin/model/PhoneEntity.java
@@ -1,4 +1,4 @@
-package com.simonalong.mikilin.type;
+package com.simonalong.mikilin.model;
 
 import com.simonalong.mikilin.annotation.FieldWhiteMatcher;
 import com.simonalong.mikilin.match.FieldModel;
@@ -8,14 +8,14 @@ import lombok.experimental.Accessors;
 
 /**
  * @author zhouzhenyong
- * @since 2019/3/10 下午9:50
+ * @since 2019/3/10 下午9:49
  */
 @Data
 @Accessors(chain = true)
-public class FixPhoneEntity {
+public class PhoneEntity {
 
-    @FieldWhiteMatcher(type = FieldModel.FIXED_PHONE)
-    private String fixedPhone;
-    @FieldBlackMatcher(model = FieldModel.FIXED_PHONE)
-    private String fixedPhoneInValid;
+    @FieldWhiteMatcher(model = FieldModel.PHONE_NUM)
+    private String phoneValid;
+    @FieldBlackMatcher(model = FieldModel.PHONE_NUM)
+    private String phoneInValid;
 }
